allow user to exclude stats for each tube in metrics
The use case required to avoid an export of stats for each tube. It is needed when there is only interest in common stats and number of tubes can significantly increase metrics size which could helpen with a large number of tubes (e.g. over 500 tubes can produced for each type of statistic quite a lot of data and it will slow to collect and agregate them from several export instances)